renamed routes init file linted code updated syntax
