+++ .gitignore
@@ -1,3 +1,6 @@
+# Database
+db.sqlite 
+
 # Logs
 logs
 *.log
